refactor(policy): split GitLab discovery into policy/_gitlab.py
discovery.py grew to 2322 lines, over the 2100-line CI file-length
ratchet. Move only this PR's new GitLab code (repo candidates, root-group
resolution, Repository Files API transport, repo fetch orchestration) into
a private policy/_gitlab.py module; pre-existing code stays put.

- discovery.py accesses the moved helpers via ``from . import _gitlab``
  (no name re-exports, no circular import; _gitlab defers its discovery
  imports to call time, matching the existing AuthResolver pattern)
- _auto_discover resolves the GitLab root group inline at the call site
- new GitLab tests patch apm_cli.policy._gitlab.* targets; all other
  patch targets and imports unchanged
